Removed LineTest, added others
Removed LineTest.java since we can't test gui stuff. Added a ViewTemplateTest and a ControllerTest to bring code coverage up, and shuffled some folders/packages around to be clearer.
