+++ src/test/java/mike/view/CLIViewTest.java
@@ -1,4 +1,4 @@
-package mike.cli;
+package mike.view;
 
 import static org.junit.Assert.*;
 
@@ -9,8 +9,6 @@ import java.io.PrintStream;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-
-import mike.view.CLIView;
 
 public class CLIViewTest {
